### 2021-07-25 #3535 @elbeejay
Hi @jessecambon, thank you for submitting `tidygeocoder` to JOSS. Per the top post in this thread:

> if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).

In the meantime I'll have a pre-review look through the repository to see if there are any changes that should be made prior to sending this out for review.

Feel free to reach out with any questions.

-Jay
