### 2021-07-25 #3535 @jessecambon
Thanks @elbeejay. Here are few potential reviewers (they have R listed as a primary language and spatial listed as a domain):

nickbearman ateucher yongzesong GISerDaishaoqing nuest giskid sjsrey
