### 2021-07-27 #3535 @elbeejay
:wave: @nickbearman, @tsamsonov, @cole-brokamp, @GISerDaishaoqing would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

The submission is "tidygeocoder: An R package for geocoding." I am reaching out to you all based on your experience with R and geospatial methods.

This is a pre-review issue which is used to find reviewers. Once 3 reviewers have been found I'll start the review on a dedicated GitHub issue. At present we are asking reviewers to complete reviews in 6 weeks, although this can be extended if needed. If you are not able to review and have someone to recommend, please mention them here (in this case please mention their GitHub handle without the "@" symbol).

If you are interested in reviewing, I would recommend looking over the journal's [conflict of interest policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html?#joss-conflict-of-interest-policy) before the review process starts.

Thank you, feel free to reach out if you have any questions about the JOSS review process. Please do not feel any pressure to accept this review request if you do not have the time or do not feel comfortable reviewing this software package, we appreciate and respect our peer-reviewers' time.

-Jay
