### 2021-07-29 #3544 @whedon
[REVIEW]: tidygeocoder: An R package for geocoding
**Submitting author:** @jessecambon (<a href="http://orcid.org/0000-0001-6854-1514">Jesse Cambon</a>)
**Repository:** <a href="https://github.com/jessecambon/tidygeocoder" target ="_blank">https://github.com/jessecambon/tidygeocoder</a>
**Version:** v1.0.3

- [x] #3674
