### 2021-07-30 1d261f0
typo fix in development notes and replaced "geocoder service" with "geocoding service"
