+++ paper.bib
@@ -1,195 +1,0 @@
- 
-@Misc{ tidygeocoder:2021,
-  title = {tidygeocoder: Geocoding Made Easy},
-  author = {Jesse Cambon and Diego Hernangómez and Christopher Belanger and Daniel Possenriede},
-  year = {2021},
-  publisher = {Zenodo},
-  note = {R package version 1.0.3},
-  url = {https://CRAN.R-project.org/package=tidygeocoder},
-  doi = {10.5281/zenodo.4686074},
-} 
- 
-@Manual{ R:2021,
-  title = {R: A Language and Environment for Statistical Computing},
-  author = {{R Core Team}},
-  organization = {R Foundation for Statistical Computing},
-  address = {Vienna, Austria},
-  year = {2021},
-  url = {https://www.R-project.org/},
-}
-
-@article{ Wickham:2014,
-   author = {Hadley  Wickham},
-   title = {Tidy Data},
-   journal = {Journal of Statistical Software, Articles},
-   volume = {59},
-   number = {10},
-   year = {2014},
-   issn = {1548-7660},
-   pages = {1--23},
-   doi = {10.18637/jss.v059.i10},
-   url = {https://www.jstatsoft.org/v059/i10}
-}
-
-@article{ Wickham:2019,
-  doi = {10.21105/joss.01686},
-  url = {https://doi.org/10.21105/joss.01686},
-  year = {2019},
-  publisher = {The Open Journal},
-  volume = {4},
-  number = {43},
-  pages = {1686},
-  author = {Hadley Wickham and Mara Averick and Jennifer Bryan and Winston Chang and Lucy D'Agostino McGowan and Romain François and Garrett Grolemund and Alex Hayes and Lionel Henry and Jim Hester and Max Kuhn and Thomas Lin Pedersen and Evan Miller and Stephan Milton Bache and Kirill Müller and Jeroen Ooms and David Robinson and Dana Paige Seidel and Vitalie Spinu and Kohske Takahashi and Davis Vaughan and Claus Wilke and Kara Woo and Hiroaki Yutani},
-  title = {Welcome to the Tidyverse},
-  journal = {Journal of Open Source Software}
-}
-
-@Manual{ httr,
-  title = {httr: Tools for Working with URLs and HTTP},
-  author = {Hadley Wickham},
-  year = {2020},
-  note = {R package version 1.4.2},
-  url = {https://CRAN.R-project.org/package=httr},
-}
-
-@Article{ jsonlite,
-  title = {The jsonlite Package: A Practical and Consistent Mapping Between JSON Data and R Objects},
-  author = {Jeroen Ooms},
-  journal = {arXiv:1403.2805 [stat.CO]},
-  year = {2014},
-  url = {https://arxiv.org/abs/1403.2805},
-}
-
-@Manual{ tibble,
-  title = {tibble: Simple Data Frames},
-  author = {Kirill Müller and Hadley Wickham},
-  year = {2021},
-  note = {R package version 3.0.6},
-  url = {https://CRAN.R-project.org/package=tibble},
-}
-
-@Manual{ dplyr,
-  title = {dplyr: A Grammar of Data Manipulation},
-  author = {Hadley Wickham and Romain François and Lionel Henry and Kirill Müller},
-  year = {2021},
-  note = {R package version 1.0.4},
-  url = {https://CRAN.R-project.org/package=dplyr},
-}
-
-@article{ Walming:2021,
-  doi = {10.1007/s13187-020-01951-7},
-  url = {https://doi.org/10.1007/s13187-020-01951-7},
-  year = {2021},
-  month = jan,
-  publisher = {Springer Science and Business Media {LLC}},
[…]
