+++ R/geocode.R
@@ -2,7 +2,7 @@
 #'
 #' @description
 #' Takes a dataframe containing addresses as an input and returns 
-#' the results from a specified geocoder service in a dataframe format using the
+#' the results from a specified geocoding service in a dataframe format using the
 #' [geo] function. See example usage in `vignette("tidygeocoder")`.
 #' 
 #' This function passes all additional parameters (`...`) to the 

[…]
