### 2021-08-16 #3544 @elbeejay
:wave: @cole-brokamp @tsamsonov we are nearing the halfway point in the review process so I just wanted to check in with you both.

Let me know if there are any questions about conducting a JOSS review and I'll do my best to provide clarification. Thanks again for agreeing to review for JOSS.
