### 2021-08-16 jessecambon/tidygeocoder#128 @cole-brokamp
JOSS review
Hello!  I am listing two smaller review points here for my JOSS review (https://github.com/openjournals/joss-reviews/issues/3544)

- consider combining LICENSE.md and LICENSE into one file to prevent confusion
- please add a section title called "Statement of Need"; I believe this is sufficiently described, but just need to organize it into a section with this specific title
