### 2021-08-16 #3544 @cole-brokamp
Hi @elbeejay, I've initiated my review.  These are two very minor changes that can hopefully be addressed easily.
