### 2021-08-16 jessecambon/tidygeocoder#128 @jessecambon
Hi @cole-brokamp, thanks. To your second bullet, the middle section "Challenges in Geocoding" is the statement of need. I can just change that section title to be "Statement of Need" if that would work for you.

To your first bullet, while I agree in principle, the CRAN regulations prevent this:

> Complete and ship as file LICENSE the following 2
lines (only)

>YEAR:
COPYRIGHT HOLDER:

> and specify as

> License: MIT + file LICENSE
-https://cran.r-project.org/web/licenses/MIT

You can see a more extended discussion on this topic here: https://github.com/tidyverse/tidyr/issues/262.
