### 2021-08-16 jessecambon/tidygeocoder#128 @dieghernan
Very common problem on JOSS reviews due to CRAN policy, see also https://github.com/kadyb/rgugik/issues/61
