### 2021-08-16 jessecambon/tidygeocoder#128 @cole-brokamp
Thanks for the information.  I don't think there is anything that can be done about the LICENSE files.  I think that renaming the middle section as Statement of Need would work to satisfy JOSS requirements.
