### 2021-08-16 jessecambon/tidygeocoder#128 @jessecambon
Sounds good, I just made that change:

https://github.com/jessecambon/tidygeocoder/commit/a111b67881eaafc8277fba792d564a58016726f5
