### 2021-08-22 c7cfe2f
working progress bar with progress package. geo function single address geocoding only #38
