+++ DESCRIPTION
@@ -35,7 +35,8 @@ Imports:
     dplyr,
     httr,
     jsonlite,
-    lifecycle
+    lifecycle,
+    progress
 Suggests: 
     knitr,
     rmarkdown,
