### 2021-08-23 1e8a1c1
for geo: made some console output default, added quiet parameter (to suppress this default output), added progress bar conditions #38
