### 2021-08-23 6ea1497
tweak geo query messages
