### 2021-08-25 f95c1b6
implemented progress bars and more verbose default messages regarding queries in reverse_geo()
