+++ R/reverse_geo.R
@@ -1,8 +1,3 @@
-## References
-# google: https://developers.google.com/maps/documentation/geocoding/start
-# geocodio: https://www.geocod.io/docs/#reverse-geocoding
-# osm: https://nominatim.org/release-docs/latest/api/Reverse/
-# opencage: https://opencagedata.com/api
 
 ## IMPORTANT: All new REVERSE batch geocoding functions must be added to reverse_batch_func_map
 # the reverse_geo() function references this list to find reverse batch geocoding functions (reverse_batch_geocoding.R)
@@ -14,6 +9,14 @@ reverse_batch_func_map <- list(
   mapquest = reverse_batch_mapquest,
   bing = reverse_batch_bing
 )
+
+# call geo function with progress bar
+progress_reverse_geo <- function(pb = NULL, ...) {
+  results <- reverse_geo(...)
+  if (!is.null(pb)) pb$tick()
+  return(results)
+}
+
 
 # Create API parameters for a single set of coordinates (lat, long) based on the 
 # method. Parameters are placed into the 'custom_query' variable which is a named list
@@ -101,7 +104,8 @@ get_coord_parameters <- function(custom_query, method, lat, long) {
 #' }
 #' @seealso [reverse_geocode] [api_parameter_reference] [min_time_reference] [batch_limit_reference]
 #' @export
-reverse_geo <- function(lat, long, method = 'osm', address = address, limit = 1, min_time = NULL, api_url = NULL,  
+reverse_geo <- function(lat, long, method = 'osm', address = address, limit = 1, min_time = NULL, 
+    progress_bar = show_progress_bar(), quiet = FALSE, api_url = NULL,  
     timeout = 20, mode = '',  full_results = FALSE, unique_only = FALSE, return_coords = TRUE, flatten = TRUE, 
     batch_limit = NULL, verbose = FALSE, no_query = FALSE, custom_query = list(), iq_region = 'us', geocodio_v = 1.6,
     mapbox_permanent = FALSE, here_request_id = NULL, mapquest_open = FALSE) {
@@ -115,12 +119,16 @@ reverse_geo <- function(lat, long, method = 'osm', address = address, limit = 1,
   
   # Check argument inputs
   stopifnot(is.logical(verbose), is.logical(no_query), is.logical(flatten),
-      is.logical(full_results), is.logical(unique_only),
+      is.logical(full_results), is.logical(unique_only), is.logical(progress_bar), 
+      is.logical(quiet),
       is.list(custom_query), 
       is.logical(mapbox_permanent),
       is.null(here_request_id) || is.character(here_request_id),
       is.logical(mapquest_open)
   )
+  if (quiet == TRUE && verbose == TRUE) {
+    stop("quiet and verbose cannot both be TRUE. See ?reverse_geo")
+  }
   
   # Check method argument
   # Currently census is the only method that doesn't support reverse geocoding
@@ -158,25 +166,35 @@ reverse_geo <- function(lat, long, method = 'osm', address = address, limit = 1,
   }
   
   # Exception for geocoding services that should default to single instead of batch
-  if (method %in% pkg.globals$single_first_methods && mode != 'batch' ){
+  if (method %in% pkg.globals$single_first_methods && mode != 'batch'){
     mode <- 'single'
   }
   
   # Geocode coordinates one at a time in a loop -------------------------------------------------------
   if ((num_unique_coords > 1) & ((!(method %in% names(reverse_batch_func_map))) | (mode == 'single'))) {
     
-    if (verbose == TRUE) {
-      message('Executing single coordinate geocoding...\n')
-    }
-    
     # construct arguments for a single address query
     # note that non-lat/long related fields go to the MoreArgs argument of mapply
     # since we aren't iterating through them
     single_coord_args <- c(
[…]
