### 2021-08-26 7fb66fe
utility functions for here method check and query completion messages
