### 2021-08-26 a0111bc
method check function
