### 2021-08-27 2278615
error checking tweaks and news update
