+++ NEWS.md
@@ -1,9 +1,9 @@
 # tidygeocoder (development version)
 
+- Added progress bar for single input geocoding ([#38](https://github.com/jessecambon/tidygeocoder/issues/38)). It can be disabled with `progress_bar = FALSE` (a new parameter for the `geo()` and `reverse_geo()` functions). By default progress bars are displayed if the session is interactive and the code being run is not part of an RStudio Notebook chunk or R Markdown knitting process.
+- Console messages for the number of inputs (addresses or coordinates) submitted, the geocoding service used, and the time elapsed are shown by default. These messages can be suppressed with `quiet = TRUE` (a new parameter for the `geo()` and `reverse_geo()` functions). The time the query was submitted is displayed for batch geocoding queries to help keep track of how much time has elapsed.
 - Added support for the [Geoapify](https://www.geoapify.com/) service (thanks [@dpprdan](https://github.com/dpprdan)). This service supports batch geocoding, but this capability is currently not implemented in tidygeocoder (see [#119](https://github.com/jessecambon/tidygeocoder/issues/119)).
 - Fixed a bug for Bing forward geocoding `geo()` when no results are found ([#112](https://github.com/jessecambon/tidygeocoder/issues/112)).
-- Added progress bar for single address geocoding ([#38](https://github.com/jessecambon/tidygeocoder/issues/38)). It can be disabled with `progress_bar = FALSE` (a new parameter for the `geo()` function). 
-- Console messages for the number of addresses submitted, the method used, and the time elapsed are shown by default. These messages can be suppressed with `quiet = TRUE` (a new parameter of the `geo()` function.
 
 # tidygeocoder 1.0.3
 
