### 2021-08-27 cb13014
fix global variable error
