### 2021-08-30 fca7a8b
tweaked language and added tests
