+++ man/extract_results.Rd
@@ -15,7 +15,7 @@ extract_results(
 \arguments{
 \item{method}{method name}
 
-\item{response}{content from the geocoder service (returned by the \link{query_api} function)}
+\item{response}{content from the geocoding service (returned by the \link{query_api} function)}
 
 \item{full_results}{if TRUE then the full results (not just latitude and longitude)
 will be returned.}
@@ -25,7 +25,7 @@ will be returned.}
 \item{limit}{only used for "census" and "google" methods. Limits number of results per address.}
 }
 \value{
-geocoder results in tibble format
+geocoding results in tibble format
 }
 \description{
 Parses the output of the \link{query_api} function for single
