+++ man/geocode.Rd
@@ -81,7 +81,7 @@ both acceptable).
 }
 \examples{
 \donttest{
-library(dplyr)
+library(dplyr, warn.conflicts = FALSE)
 sample_addresses \%>\% slice(1:2) \%>\%
  geocode(addr, method = 'arcgis')
 
