+++ man/reverse_geocode.Rd
@@ -58,7 +58,7 @@ on possible arguments.
 \examples{
 \donttest{
 library(tibble)
-library(dplyr)
+library(dplyr, warn.conflicts = FALSE)
 
 tibble(
     latitude = c(38.895865, 43.6534817),
@@ -77,7 +77,7 @@ louisville \%>\% head(3) \%>\%
 
 louisville \%>\% head(2) \%>\% 
   reverse_geocode(lat = latitude, long = longitude,  
-  method = 'osm', verbose = TRUE,
+  method = 'osm',
   limit = 2, return_input = FALSE)
 
 }
