### 2021-08-31 a0d6c67
fix for query message not being displayed when n=1
