+++ man/reverse_geo.Rd
@@ -28,7 +28,8 @@ reverse_geo(
   geocodio_v = 1.6,
   mapbox_permanent = FALSE,
   here_request_id = NULL,
-  mapquest_open = FALSE
+  mapquest_open = FALSE,
+  init = TRUE
 )
 }
 \arguments{
@@ -124,6 +125,8 @@ needs to be FALSE.}
 
 \item{mapquest_open}{if TRUE then MapQuest would use the Open Geocoding
 endpoint, that relies solely on data contributed to OpenStreetMap.}
+
+\item{init}{internal use only. Set to TRUE for initial query and FALSE otherwise.}
 }
 \value{
 tibble (dataframe)
