### 2021-08-31 jessecambon/tidygeocoder#133 @tsamsonov
Collapse `method` and `cascade _order` parameters to a single `service` parameter in `geo()` function
Dear @jessecambon and co-authors,

Thanks for this great package. I am currently reviewing the corresponding JOSS paper (https://github.com/openjournals/joss-reviews/issues/3544)

I would like to discuss the possible improvement of the `geo()` function interface.

Two thoughts:

1. `method` is a poor name for a parameter which selects the _service_ used for geocoding. From a scientific point of view, the method is the actual formalized algorithmic procedure with math background that is used to convert an address into coordinates. Since most of the geocoders are actually proprietary, we do not know what method of geocoding is really used under the hood. Since your package actually interacts with geocoder APIs, it is more suitable to name this parameter `service`. This is clear and reflects its real dedication.
2. If you make the possibility to pass the vector with one (actually scalar) or two character values to the `service` parameter, then you do not need an additional `cascade_order ` parameter. If the user passes a single service name, then use the standard approach. If the vector is of the length two, then use the cascade approach.

I am looking forward for your opinion on these suggestions.

All the best, Tim
