### 2021-09-01 jessecambon/tidygeocoder#133 @jessecambon
Hi @tsamsonov, thanks for your review and suggestions. I'll give my thoughts and @dpprdan @dieghernan @chris31415926535 can feel free to weigh in as well.

Regarding your first suggestion, my thinking behind naming the argument `method` was that (1) from the user's functional perspective, the geocoding service is the tool/method for accomplishing the task of geocoding and (2) multiple services can be used with the same value of the `method` argument so it isn't necessarily selecting the service.

For instance, while by default `method = 'osm'` uses the publicly available Nominatim service hosted by the OpenStreetMap project, you could also choose to host your own Nominatim service on a private server since Nominatim is a freely available open source product. You can specify the URL for your service with the `api_url` argument and still use `method = 'osm'`. One of the supported geocoding services, [Location IQ](https://locationiq.com/), is heavily based on Nominatim to the point that the only difference between the "osm" and "iq" methods in the tidygeocoder code that I'm aware of is the use of an API key in queries. [Pelias](https://pelias.io/) is another open source geocoder that isn't currently supported by the package, but could be in the future.

For your second suggestion, I think that using a character vector for `method` (instead of `method = "cascade"`) and deprecating the `cascade_order` argument like you suggest could potentially be a good approach that wouldn't be too much of a headache to implement. This would only be required for geo() since cascade was never implemented for reverse geocoding. However, it might be worth also considering if we even want to keep the 'cascade' method. As the package has grown I've considered deprecating the "cascade" method and mainly kept it around to avoid breaking existing code.
