### 2021-09-01 jessecambon/tidygeocoder#133 @jessecambon
See https://github.com/jessecambon/tidygeocoder/issues/134#issuecomment-910747448 for a proposal for a new cascade function that could replace the current cascade method.
