### 2021-09-01 f9284f7
made utility function for api url modifications
