### 2021-09-02 #3544 @jessecambon
We updated the paper to include a comparison to other R packages (state of the field).
