### 2021-09-04 0b0c174
add default arguments in options()
