+++ R/utils.R
@@ -262,8 +262,8 @@ extract_bing_latlng <- function(response) {
 # For consistency/continuity, these are the same conditions 
 # that are used for the {readr} package
 show_progress_bar <- function() {
- # isTRUE(getOption("tidygeocoder.show_progress")) && # user disables progress bar
-    interactive() && # an interactive session
+    isTRUE(getOption("tidygeocoder.progress_bar")) && # options() setting isn't FALSE
+    interactive() && # interactive session
     !isTRUE(getOption("rstudio.notebook.executing")) && # Not running in an RStudio notebook chunk
     !isTRUE(getOption("knitr.in.progress")) # Not actively knitting a document
 }
