+++ man/geo.Rd
@@ -19,17 +19,17 @@ geo(
   limit = 1,
   min_time = NULL,
   progress_bar = show_progress_bar(),
-  quiet = FALSE,
+  quiet = isTRUE(getOption("tidygeocoder.quiet")),
   api_url = NULL,
   timeout = 20,
   mode = "",
-  full_results = FALSE,
+  full_results = isTRUE(getOption("tidygeocoder.full_results")),
   unique_only = FALSE,
   return_addresses = TRUE,
-  flatten = TRUE,
+  flatten = isTRUE(getOption("tidygeocoder.flatten")),
   batch_limit = NULL,
   batch_limit_error = TRUE,
-  verbose = FALSE,
+  verbose = isTRUE(getOption("tidygeocoder.verbose")),
   no_query = FALSE,
   custom_query = list(),
   return_type = "locations",
@@ -117,8 +117,8 @@ be used if available (given method selected) when multiple addresses are
 provided; otherwise single address geocoding will be used. For the "here" and "bing" methods the
 batch mode should be explicitly specified with \code{mode = 'batch'}.}
 
-\item{full_results}{returns all data from the geocoding service if TRUE.
-If FALSE then only longitude and latitude are returned from the geocoding service.}
+\item{full_results}{returns all available data from the geocoding service if TRUE.
+If FALSE (default) then only latitude and longitude columns are returned from the geocoding service.}
 
 \item{unique_only}{only return results for unique inputs if TRUE}
 
@@ -126,7 +126,7 @@ If FALSE then only longitude and latitude are returned from the geocoding servic
 most services return the input addresses with \code{full_results = TRUE} and setting
 return_addresses to FALSE does not prevent this.}
 
-\item{flatten}{if TRUE then any nested dataframes in results are flattened if possible.
+\item{flatten}{if TRUE (default) then any nested dataframes in results are flattened if possible.
 Note that in some cases results are flattened regardless such as for
 Geocodio batch geocoding.}
 
@@ -136,7 +136,7 @@ Defaults to the value in \link{batch_limit_reference} if not specified.}
 \item{batch_limit_error}{if TRUE then an error is thrown if the number of addresses exceeds the batch limit.
 (if executing a batch query). This is reverted to FALSE when using the cascade method.}
 
-\item{verbose}{if TRUE then detailed logs are output to the console}
+\item{verbose}{if TRUE then detailed logs are output to the console. FALSE is default.}
 
 \item{no_query}{if TRUE then no queries are sent to the geocoding service and verbose is set to TRUE.
 Used for testing.}

[…]
