### 2021-09-04 jessecambon/tidygeocoder#133 @tsamsonov
Hi @jessecambon!

Considering the suggestion about the dedicated cascade function — I think it is a reasonable solution. It will potentially bring more flexibility in combining multiple services, while keeping the core single-service functionality in `geo()` function. In light of this change I have two suggestions. First, it can be reasonable to select the order of geocoding in automated way. For example, the first service geocodes only 30% of addresses, and the remaining 70% are successfully geocoded by the second service. But if you execute a query against the second service first, then it may appear, that this service geocodes all 100% addresses successfully, and there is no need to use the first service at all. My second suggestion is to remove the limit of two services. I can imagine the situation when the user does not know what service will bring the best results. Then it will convenient to have something like `geo_cascade(df, service = 'all')` and let the function select the best geocoding service. Also, the user may potentially want to use an arbitrary list of 3, 5 or other number of services, not necessarily 1 or 2. I understand that batch geocoding can be time consuming, and such a "geocoder roundabout" will bring a lot of computational overhead. But at the same time, this will be the user's choice.

As for the naming of the `method` parameter — I understand your reasoning. However, I still think that `service` is a better name. For example, there is no contradiction in using the `service = osm`, and `api_url = localhost\...`, since in this case the _service_ is local. Btw, why not `service = nominatim`? Or you call the services by their developer's name?

Al the best, Tim
