### 2021-09-04 jessecambon/tidygeocoder#133 @jessecambon
Hi Tim, I agree regarding allowing more than two services to be used for the new cascade function. I'm thinking letting the user pass a lists of lists could accomplish this (ie. `cascade_function(list(list(<query 1 args>), list(query 2 args>), list(<query 3 args>), ...))`. The two freely available services, ArcGIS and Nominatim, could be defaults for the function. Probably with ArcGIS first, because Nominatim has greater restrictions on the rate of querying...

We can give some thought to the name of the `method` argument, but I would just want to make sure changing the name is worth the hassle for users to change their code. `geocoder` could be an appropriate name for that argument as well, but it might be a bit redundant in a function named "geocode" that's in a package called "tidygeocoder"...

As for your last question, I used `method = 'osm'` instead of `method = 'nominatim'` just because it's faster to type and the spelling is easier to remember.
