### 2021-09-07 #3544 @elbeejay
Hi @jessecambon and coauthors, I'm pleased to see that the reviewers have completed their review of your package. I have a few minor clarifications and changes of my own to request.

- As you can see in the reference check above, whedon has flagged the DOI for `ggmap` as missing. Please check that you are correctly citing that package. There appear to be multiple accepted ways of citing it: [method 1](https://journal.r-project.org/archive/2013/RJ-2013-014/index.html) and [method 2](https://cran.r-project.org/web/packages/ggmap/citation.html). I suggest using the method with the DOI suggested by whedon above.
- I noticed that the first sentence of the paper has a reference to a Zenodo archive of this software. Please remove this citation as the JOSS paper will serve as a connection to the published version of `tidygeocoder` and will include metadata and a link to a tagged version in Zenodo (more on this below).

Once you have made the above changes, there are just a few things I need from you to set the metadata and mark this as "ready for publication"; please:

Given that the software (as far as I am aware) did not actually change as a part of this review, these steps may already be complete. I just ask that you double-check the metadata to ensure everything is correct, and then list the requested items here. I will then move forward with accepting the submission, thanks!
