### 2021-09-07 #3544 @jessecambon
Hi @elbeejay, thanks. I've made those two changes and the Zenodo archive is here : https://zenodo.org/record/4686074#.YTdlIfxKiV5 (github release v1.0.3). I double checked Zenodo metadata and it is correct.
